+++ apps/web/src/utils/helpers.ts
@@ -261,3 +261,10 @@ export function unpackOptionalArray<T = []>(
   }
   return array
 }
+
+export function maxChar(str: string, maxLength: number) {
+  if (str.length <= maxLength) {
+    return str
+  }
+  return str.slice(0, maxLength) + '...'
+}
